sample class
