+++ sim/utils.py
@@ -1,0 +1,21 @@
+import subprocess
+
+def run(cmd, *args, **kwargs):
+    """ Run a command and assert that the process exits with a non-zero exit code.
+        See python's subprocess.run command for args/kwargs
+
+        Parameters:
+            cmd (list): List of strings defining the command, see (subprocess.run in python docs)
+            cwd (str): Set surr
+
+        Returns:
+            None
+    """
+    # TODO: store stdout to a file
+    returncode = subprocess.run(cmd, *args, **kwargs).returncode
+
+    if returncode:
+        raise Exception("""*****
+            %s
+            cmd failed with exit code %i
+          *****""" % (cmd, returncode))
